requirements.txt
fastapi
uvicorn
tensorflow
opencv-python-headless
numpy
scikit-learn
matplotlib
pandas
streamlit
requests
pillow
numpy
pandas
python-multipart

